requirements.txt
fastapi
uvicorn[standard]
httpx
reportlab
base58
python-dotenv
sqlalchemy
